+++ processor/src/operations/mod.rs
@@ -6,6 +6,7 @@ use vm_core::{
 use super::{ExecutionError, Felt, FieldElement, Host, Operation, Process};
 use crate::errors::ErrorContext;
 
+mod circuit_eval;
 mod crypto_ops;
 mod ext2_ops;
 mod field_ops;
@@ -180,6 +181,7 @@ impl Process {
             Operation::FriE2F4 => self.op_fri_ext2fold4()?,
             Operation::HornerBase => self.op_horner_eval_base(error_ctx)?,
             Operation::HornerExt => self.op_horner_eval_ext(error_ctx)?,
+            Operation::ArithmeticCircuitEval => self.arithmetic_circuit_eval(error_ctx)?,
         }
 
         self.advance_clock()?;